test: split shortcuts tests
